Moved the flow and some reusable functions of liquidity_wall into advanced_strategy.py
